Update hoard-screen.component.html
I want to use the index to target individual items even if they are duplicates. in the final version duplicates will be far less likely (I hope) but this is still a problem. I can't just not allow duplicates!
note: if i remove coins from the rng options, i can leave coins as the 5% or so remainder after other items generated. the problem with that is low-value encounter budgets